+++ test.py
@@ -6,12 +6,13 @@
 #
 # SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.0
 #
+
+from __future__ import annotations
+
 import argparse
 import asyncio
 import collections
 import colorama
-import difflib
-import filecmp
 import glob
 import itertools
 import logging
@@ -20,1258 +21,35 @@ import os
 import pathlib
 import re
 import resource
-import shlex
 import shutil
 import signal
-import socket
 import subprocess
 import sys
 import time
-import traceback
 import xml.etree.ElementTree as ET
-import yaml
-from random import randint
-from tempfile import TemporaryDirectory
+from typing import TYPE_CHECKING
 
-from abc import ABC, abstractmethod
-from io import StringIO
-from scripts import coverage    # type: ignore
-from test.pylib.artifact_registry import ArtifactRegistry
-from test.pylib.cpp.ldap.prepare_instance import try_something_backoff, can_connect, setup
-from test.pylib.host_registry import HostRegistry
-from test.pylib.pool import Pool
-from test.pylib.s3_proxy import S3ProxyServer
-from test.pylib.s3_server_mock import MockS3Server
-from test.pylib.resource_gather import setup_cgroup, run_resource_watcher, get_resource_gather
-from test.pylib.util import LogPrefixAdapter, get_configured_modes, ninja
-from test.pylib.scylla_cluster import ScyllaServer, ScyllaCluster, get_cluster_manager, merge_cmdline_options
-from test.pylib.minio_server import MinioServer
-from typing import Dict, List, Callable, Any, Iterable, Optional, Awaitable, Union
-import logging
-from test.pylib import coverage_utils
 import humanfriendly
 import treelib
 
+from scripts import coverage
+from test.pylib import coverage_utils
+from test.pylib.cpp.ldap.prepare_instance import try_something_backoff, can_connect
+from test.pylib.host_registry import HostRegistry
+from test.pylib.minio_server import MinioServer
+from test.pylib.s3_proxy import S3ProxyServer
+from test.pylib.s3_server_mock import MockS3Server
+from test.pylib.suite.base import Test, TestSuite, all_modes, output_is_a_tty, palette, path_to
+from test.pylib.suite.boost import BoostTest
+from test.pylib.suite.ldap import LdapTest
+from test.pylib.resource_gather import setup_cgroup, run_resource_watcher
+from test.pylib.util import LogPrefixAdapter, get_configured_modes, ninja
+
+if TYPE_CHECKING:
+    from typing import List
+
+
 launch_time = time.monotonic()
-
-output_is_a_tty = sys.stdout.isatty()
-
-all_modes = {'debug': 'Debug',
-             'release': 'RelWithDebInfo',
-             'dev': 'Dev',
-             'sanitize': 'Sanitize',
-             'coverage': 'Coverage'}
-debug_modes = {'debug', 'sanitize'}
-
-
-def create_formatter(*decorators) -> Callable[[Any], str]:
-    """Return a function which decorates its argument with the given
-    color/style if stdout is a tty, and leaves intact otherwise."""
-    def color(arg: Any) -> str:
-        return "".join(decorators) + str(arg) + colorama.Style.RESET_ALL
-
-    def nocolor(arg: Any) -> str:
-        return str(arg)
-    return color if output_is_a_tty else nocolor
-
-
-class palette:
-    """Color palette for formatting terminal output"""
-    ok = create_formatter(colorama.Fore.GREEN, colorama.Style.BRIGHT)
-    fail = create_formatter(colorama.Fore.RED, colorama.Style.BRIGHT)
-    new = create_formatter(colorama.Fore.BLUE)
-    skip = create_formatter(colorama.Style.DIM)
-    path = create_formatter(colorama.Style.BRIGHT)
-    diff_in = create_formatter(colorama.Fore.GREEN)
-    diff_out = create_formatter(colorama.Fore.RED)
-    diff_mark = create_formatter(colorama.Fore.MAGENTA)
-    warn = create_formatter(colorama.Fore.YELLOW)
-    crit = create_formatter(colorama.Fore.RED, colorama.Style.BRIGHT)
-    ansi_escape = re.compile(r'\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])')
-    @staticmethod
-    def nocolor(text: str) -> str:
-        return palette.ansi_escape.sub('', text)
-
-
-def path_to(mode, *components):
-    """Resolve path to built executable"""
-    build_dir = 'build'
-    if os.path.exists(os.path.join(build_dir, 'build.ninja')):
-        *dir_components, basename = components
-        return os.path.join(build_dir, *dir_components, all_modes[mode], basename)
-    return os.path.join(build_dir, mode, *components)
-
-
-class TestSuite(ABC):
-    """A test suite is a folder with tests of the same type.
-    E.g. it can be unit tests, boost tests, or CQL tests."""
-
-    # All existing test suites, one suite per path/mode.
-    suites: Dict[str, 'TestSuite'] = dict()
-    artifacts = ArtifactRegistry()
-    hosts = HostRegistry()
-    FLAKY_RETRIES = 5
-    _next_id = collections.defaultdict(int) # (test_key -> id)
-
-    def __init__(self, path: str, cfg: dict, options: argparse.Namespace, mode: str) -> None:
-        self.suite_path = pathlib.Path(path)
-        self.name = str(self.suite_path.name)
-        self.cfg = cfg
-        self.options = options
-        self.mode = mode
-        self.suite_key = os.path.join(path, mode)
-        self.tests: List['Test'] = []
-        self.pending_test_count = 0
-        # The number of failed tests
-        self.n_failed = 0
-
-        self.run_first_tests = set(cfg.get("run_first", []))
-        self.no_parallel_cases = set(cfg.get("no_parallel_cases", []))
-        # Skip tests disabled in suite.yaml
-        self.disabled_tests = set(self.cfg.get("disable", []))
-        # Skip tests disabled in specific mode.
-        self.disabled_tests.update(self.cfg.get("skip_in_" + mode, []))
-        self.flaky_tests = set(self.cfg.get("flaky", []))
-        # If this mode is one of the debug modes, and there are
-        # tests disabled in a debug mode, add these tests to the skip list.
-        if mode in debug_modes:
-            self.disabled_tests.update(self.cfg.get("skip_in_debug_modes", []))
-        # If a test is listed in run_in_<mode>, it should only be enabled in
-        # this mode. Tests not listed in any run_in_<mode> directive should
-        # run in all modes. Inversing this, we should disable all tests
-        # which are listed explicitly in some run_in_<m> where m != mode
-        # This of course may create ambiguity with skip_* settings,
-        # since the priority of the two is undefined, but oh well.
-        run_in_m = set(self.cfg.get("run_in_" + mode, []))
-        for a in all_modes:
-            if a == mode:
-                continue
-            skip_in_m = set(self.cfg.get("run_in_" + a, []))
-            self.disabled_tests.update(skip_in_m - run_in_m)
-        # environment variables that should be the base of all processes running in this suit
-        self.base_env = {}
-        if self.need_coverage():
-            # Set the coverage data from each instrumented object to use the same file (and merged into it with locking)
-            # as long as we don't need test specific coverage data, this looks sufficient. The benefit of doing this in
-            # this way is that the storage will not be bloated with coverage files (each can weigh 10s of MBs so for several
-            # thousands of tests it can easily reach 10 of GBs)
-            # ref: https://clang.llvm.org/docs/SourceBasedCodeCoverage.html#running-the-instrumented-program
-            self.base_env["LLVM_PROFILE_FILE"] = os.path.join(options.tmpdir,self.mode, "coverage", self.name, "%m.profraw")
-    # Generate a unique ID for `--repeat`ed tests
-    # We want these tests to have different XML IDs so test result
-    # processors (Jenkins) don't merge results for different iterations of
-    # the same test. We also don't want the ids to be too random, because then
-    # there is no correlation between test identifiers across multiple
-    # runs of test.py, and so it's hard to understand failure trends. The
-    # compromise is for next_id() results to be unique only within a particular
-    # test case. That is, we'll have a.1, a.2, a.3, b.1, b.2, b.3 rather than
-    # a.1 a.2 a.3 b.4 b.5 b.6.
-    def next_id(self, test_key) -> int:
-        TestSuite._next_id[test_key] += 1
-        return TestSuite._next_id[test_key]
-
-    @staticmethod
-    def test_count() -> int:
-        return sum(TestSuite._next_id.values())
-
-    @staticmethod
-    def load_cfg(path: str) -> dict:
-        with open(os.path.join(path, "suite.yaml"), "r") as cfg_file:
-            cfg = yaml.safe_load(cfg_file.read())
-            if not isinstance(cfg, dict):
-                raise RuntimeError("Failed to load tests in {}: suite.yaml is empty".format(path))
-            return cfg
-
-    @staticmethod
-    def opt_create(path: str, options: argparse.Namespace, mode: str) -> 'TestSuite':
-        """Return a subclass of TestSuite with name cfg["type"].title + TestSuite.
-        Ensures there is only one suite instance per path."""
-        suite_key = os.path.join(path, mode)
-        suite = TestSuite.suites.get(suite_key)
-        if not suite:
-            cfg = TestSuite.load_cfg(path)
-            kind = cfg.get("type")
-            if kind is None:
-                raise RuntimeError("Failed to load tests in {}: suite.yaml has no suite type".format(path))
-
-            def suite_type_to_class_name(suite_type: str) -> str:
-                if suite_type.casefold() == "Approval".casefold():
-                    suite_type = "CQLApproval"
-                else:
-                    suite_type = suite_type.title()
-                return suite_type + "TestSuite"
-
-            SpecificTestSuite = globals().get(suite_type_to_class_name(kind))
-            if not SpecificTestSuite:
-                raise RuntimeError("Failed to load tests in {}: suite type '{}' not found".format(path, kind))
-            suite = SpecificTestSuite(path, cfg, options, mode)
-            assert suite is not None
-            TestSuite.suites[suite_key] = suite
-        return suite
-
-    @staticmethod
-    def all_tests() -> Iterable['Test']:
-        return itertools.chain(*(suite.tests for suite in
-                                 TestSuite.suites.values()))
-
-    @property
-    @abstractmethod
-    def pattern(self) -> str:
-        pass
-
-    @abstractmethod
-    async def add_test(self, shortname: str, casename: str) -> None:
-        pass
-
-    async def run(self, test: 'Test', options: argparse.Namespace):
-        try:
-            test.started = True
-            for i in range(1, self.FLAKY_RETRIES):
-                if i > 1:
-                    test.is_flaky_failure = True
-                    logging.info("Retrying test %s after a flaky fail, retry %d", test.uname, i)
-                    test.reset()
-                await test.run(options)
-                if test.success or not test.is_flaky or test.is_cancelled:
-                    break
-        except asyncio.CancelledError:
-            test.is_cancelled = True
-            raise
-        finally:
-            self.pending_test_count -= 1
-            self.n_failed += int(test.failed)
-            if self.pending_test_count == 0:
-                await TestSuite.artifacts.cleanup_after_suite(self, self.n_failed > 0)
-        return test
-
-    def junit_tests(self):
-        """Tests which participate in a consolidated junit report"""
-        return self.tests
-
-    def boost_tests(self):
-        return []
-
-    def build_test_list(self) -> List[str]:
-        pattern = self.pattern if isinstance(self.pattern, list) else [self.pattern]
-        tests = itertools.chain(*[self.suite_path.rglob(i) for i in pattern])
-        return [os.path.splitext(t.relative_to(self.suite_path))[0] for t in tests]
-
-    async def add_test_list(self) -> None:
-        options = self.options
-        lst = self.build_test_list()
-        if lst:
-            # Some tests are long and are better to be started earlier,
-            # so pop them up while sorting the list
-            lst.sort(key=lambda x: (x not in self.run_first_tests, x))
-
-        pending = set()
-        for shortname in lst:
-            testname = os.path.join(self.name, shortname)
-            casename = None
-
-            # Check opt-out lists
-            if shortname in self.disabled_tests:
-                continue
-            if options.skip_patterns:
-                if any(skip_pattern in testname for skip_pattern in options.skip_patterns):
-                    continue
-
-            # Check opt-in list
-            if options.name:
-                for p in options.name:
-                    pn = p.split('::', 2)
-                    if len(pn) == 1 and p in testname:
-                        break
-                    if len(pn) == 2 and pn[0] == testname:
-                        if pn[1] != "*":
-                            casename = pn[1]
-                        break
-                else:
-                    continue
-
-            async def add_test(shortname, casename) -> None:
-                # Add variants of the same test sequentially
-                # so that case cache has a chance to populate
-                for i in range(options.repeat):
-                    await self.add_test(shortname, casename)
-                    self.pending_test_count += 1
-
-            pending.add(asyncio.create_task(add_test(shortname, casename)))
-
-        if len(pending) == 0:
-            return
-        try:
-            await asyncio.gather(*pending)
-        except asyncio.CancelledError:
-            for task in pending:
-                task.cancel()
-            await asyncio.gather(*pending, return_exceptions=True)
-            raise
-    def need_coverage(self):
-        return self.options.coverage and (self.mode in self.options.coverage_modes) and bool(self.cfg.get("coverage",True))
-
-class UnitTestSuite(TestSuite):
-    """TestSuite instantiation for non-boost unit tests"""
-
-    def __init__(self, path: str, cfg: dict, options: argparse.Namespace, mode: str) -> None:
-        super().__init__(path, cfg, options, mode)
-        # Map of custom test command line arguments, if configured
-        self.custom_args = cfg.get("custom_args", {})
-        self.extra_cmdline_options = cfg.get("extra_scylla_cmdline_options", [])
-        # Map of tests that cannot run with compaction groups
-        self.all_can_run_compaction_groups_except = cfg.get("all_can_run_compaction_groups_except")
-
-    async def create_test(self, shortname, casename, suite, args):
-        exe = path_to(suite.mode, "test", suite.name, shortname)
-        if not os.access(exe, os.X_OK):
-            print(palette.warn(f"Unit test executable {exe} not found."))
-            return
-        test = UnitTest(self.next_id((shortname, self.suite_key)), shortname, suite, args)
-        self.tests.append(test)
-
-    def prepare_arg(self, arg):
-        extra_cmdline_options = ' '.join(self.extra_cmdline_options)
-        return f'{arg} {extra_cmdline_options}'
-
-    async def add_test(self, shortname, casename) -> None:
-        """Create a UnitTest class with possibly custom command line
-        arguments and add it to the list of tests"""
-        # Skip tests which are not configured, and hence are not built
-        if os.path.join("test", self.name, shortname) not in self.options.tests:
-            return
-
-        # Default seastar arguments, if not provided in custom test options,
-        # are two cores and 2G of RAM
-        args = self.custom_args.get(shortname, ["-c2 -m2G"])
-        args = merge_cmdline_options(args, self.options.extra_scylla_cmdline_options)
-        for a in args:
-            await self.create_test(shortname, casename, self, self.prepare_arg(a))
-
-    @property
-    def pattern(self) -> str:
-        # This should only match individual tests and not combined_tests.cc
-        # file of the combined test.
-        # It is because combined_tests.cc itself does not contain any tests.
-        # To keep the code simple, we have avoided this by renaming
-        # combined test file to “_tests.cc” instead of changing the match
-        # pattern.
-        return "*_test.cc"
-
-
-class BoostTestSuite(UnitTestSuite):
-    """TestSuite for boost unit tests"""
-
-    # A cache of individual test cases, for which we have called
-    # --list_content. Static to share across all modes.
-    _case_cache: Dict[str, List[str]] = dict()
-
-    _exec_name_cache: Dict[str, str] = dict()
-
-    def _generate_cache(self, exec_path, exec_name) -> None:
-        res = subprocess.run(
-                [exec_path, '--list_content'],
-                check=True,
-                capture_output=True,
-                env=dict(os.environ,
-                         **{"ASAN_OPTIONS": "halt_on_error=0"}),
-            )
-        testname = None
-        fqname = None
-        for line in res.stderr.decode().splitlines():
-            if not line.startswith('    '):
-                testname = line.strip().rstrip('*')
-                fqname = os.path.join(self.mode, self.name, testname)
-                self._exec_name_cache[fqname] = exec_name
-                self._case_cache[fqname] = []
-            else:
-                casename = line.strip().rstrip('*')
-                if casename.startswith('_'):
-                    continue
-                self._case_cache[fqname].append(casename)
-
-    def __init__(self, path, cfg: dict, options: argparse.Namespace, mode) -> None:
-        super().__init__(path, cfg, options, mode)
-        exec_name = 'combined_tests'
-        exec_path = path_to(self.mode, "test", self.name, exec_name)
-        # Apply combined test only for test/boost,
-        # cache the tests only if the executable exists, so we can
-        # run test.py with a partially built tree
-        if self.name == 'boost' and os.path.exists(exec_path):
-            self._generate_cache(exec_path, exec_name)
-
-    async def create_test(self, shortname: str, casename: str, suite, args) -> None:
-        fqname = os.path.join(self.mode, self.name, shortname)
-        if fqname in self._exec_name_cache:
-            execname = self._exec_name_cache[fqname]
-            combined_test = True
-        else:
-            execname = None
-            combined_test = False
-        exe = path_to(suite.mode, "test", suite.name, execname if combined_test else shortname)
-        if not os.access(exe, os.X_OK):
-            print(palette.warn(f"Boost test executable {exe} not found."))
-            return
-        options = self.options
-        allows_compaction_groups = self.all_can_run_compaction_groups_except != None and shortname not in self.all_can_run_compaction_groups_except
-        if options.parallel_cases and (shortname not in self.no_parallel_cases) and casename is None:
-            fqname = os.path.join(self.mode, self.name, shortname)
-            # since combined tests are preloaded to self._case_cache, this will
-            # only run in non-combined test mode
-            if fqname not in self._case_cache:
-                process = await asyncio.create_subprocess_exec(
-                    exe, *['--list_content'],
-                    stderr=asyncio.subprocess.PIPE,
-                    stdout=asyncio.subprocess.PIPE,
-                    env=dict(os.environ,
-                             **{"ASAN_OPTIONS": "halt_on_error=0"}),
-                    preexec_fn=os.setsid,
-                )
-                _, stderr = await asyncio.wait_for(process.communicate(), options.timeout)
-                # --list_content produces the list of all test cases in the file. When BOOST_DATA_TEST_CASE is used it
-                # will additionally produce the lines with numbers for each case preserving the function name like this:
-                # test_singular_tree_ptr_sz*
-                #     _0*
-                #     _1*
-                #     _2*
-                # however, it's only possible to run test_singular_tree_ptr_sz that will execute all test cases
-                # this line catches only test function name ignoring unrelated lines like '_0'
-                # Note: this will ignore any test case starting with a '_' symbol
-                case_list = [case[:-1] for case in stderr.decode().splitlines() if case.endswith('*') and not case.strip().startswith('_')]
-                self._case_cache[fqname] = case_list
-
-            case_list = self._case_cache[fqname]
-            if len(case_list) == 1:
-                test = BoostTest(self.next_id((shortname, self.suite_key)), shortname, suite, args, None, allows_compaction_groups, execname)
-                self.tests.append(test)
-            else:
-                for case in case_list:
-                    test = BoostTest(self.next_id((shortname, self.suite_key, case)), shortname, suite, args, case, allows_compaction_groups, execname)
-                    self.tests.append(test)
-        else:
-            test = BoostTest(self.next_id((shortname, self.suite_key)), shortname, suite, args, casename, allows_compaction_groups, execname)
-            self.tests.append(test)
-
-    async def add_test(self, shortname, casename) -> None:
-        """Create a UnitTest class with possibly custom command line
-        arguments and add it to the list of tests"""
-        fqname = os.path.join(self.mode, self.name, shortname)
-        if fqname in self._exec_name_cache:
-            execname = self._exec_name_cache[fqname]
-            combined_test = True
-        else:
-            combined_test = False
-        # Skip tests which are not configured, and hence are not built
-        if os.path.join("test", self.name, execname if combined_test else shortname) not in self.options.tests:
-                return
-        # Default seastar arguments, if not provided in custom test options,
-        # are two cores and 2G of RAM
-        args = self.custom_args.get(shortname, ["-c2 -m2G"])
-        args = merge_cmdline_options(args, self.options.extra_scylla_cmdline_options)
-        for a in args:
-            await self.create_test(shortname, casename, self, self.prepare_arg(a))
-
-    def junit_tests(self) -> Iterable['Test']:
-        """Boost tests produce an own XML output, so are not included in a junit report"""
-        return []
-
-    def boost_tests(self) -> Iterable['Tests']:
-        return self.tests
-
-
-class LdapTestSuite(UnitTestSuite):
-    """TestSuite for ldap unit tests"""
-
-    async def create_test(self, shortname, casename, suite, args):
-        test = LdapTest(self.next_id((shortname, self.suite_key)), shortname, suite, args)
-        self.tests.append(test)
-
-    def junit_tests(self):
-        """Ldap tests produce an own XML output, so are not included in a junit report"""
-        return []
-
-
-class PythonTestSuite(TestSuite):
-    """A collection of Python pytests against a single Scylla instance"""
-
-    def __init__(self, path, cfg: dict, options: argparse.Namespace, mode: str) -> None:
-        super().__init__(path, cfg, options, mode)
-        self.scylla_exe = path_to(self.mode, "scylla")
-        self.scylla_env = dict(self.base_env)
-        if self.mode == "coverage":
-            self.scylla_env.update(coverage.env(self.scylla_exe, distinct_id=self.name))
-        self.scylla_env['SCYLLA'] = self.scylla_exe
-
-        cluster_cfg = self.cfg.get("cluster", {"initial_size": 1})
-        cluster_size = cluster_cfg["initial_size"]
-        env_pool_size = os.getenv("CLUSTER_POOL_SIZE")
-        if options.cluster_pool_size is not None:
-            pool_size = options.cluster_pool_size
-        elif env_pool_size is not None:
-            pool_size = int(env_pool_size)
-        else:
-            pool_size = cfg.get("pool_size", 2)
-        self.dirties_cluster = set(cfg.get("dirties_cluster", []))
-
-        self.create_cluster = self.get_cluster_factory(cluster_size, options)
-        async def recycle_cluster(cluster: ScyllaCluster) -> None:
-            """When a dirty cluster is returned to the cluster pool,
-               stop it and release the used IPs. We don't necessarily uninstall() it yet,
-               which would delete the log file and directory - we might want to preserve
-               these if it came from a failed test.
-            """
-            for srv in cluster.running.values():
-                srv.log_file.close()
-                srv.maintenance_socket_dir.cleanup()
-            await cluster.stop()
-            await cluster.release_ips()
-
-        self.clusters = Pool(pool_size, self.create_cluster, recycle_cluster)
-
-    def get_cluster_factory(self, cluster_size: int, options: argparse.Namespace) -> Callable[..., Awaitable]:
-        def create_server(create_cfg: ScyllaCluster.CreateServerParams):
-            cmdline_options = self.cfg.get("extra_scylla_cmdline_options", [])
-            if type(cmdline_options) == str:
-                cmdline_options = [cmdline_options]
-            cmdline_options = merge_cmdline_options(cmdline_options, create_cfg.cmdline_from_test)
-            cmdline_options = merge_cmdline_options(cmdline_options, options.extra_scylla_cmdline_options)
-            # There are multiple sources of config options, with increasing priority
-            # (if two sources provide the same config option, the higher priority one wins):
-            # 1. the defaults
-            # 2. suite-specific config options (in "extra_scylla_config_options")
-            # 3. config options from tests (when servers are added during a test)
-            default_config_options = \
-                    {"authenticator": "PasswordAuthenticator",
-                     "authorizer": "CassandraAuthorizer"}
-            default_config_options["tablets_initial_scale_factor"] = 4 if self.mode == "release" else 2
-            config_options = default_config_options | \
-                             self.cfg.get("extra_scylla_config_options", {}) | \
-                             create_cfg.config_from_test
-
-            server = ScyllaServer(
-                mode=self.mode,
-                exe=self.scylla_exe,
-                vardir=os.path.join(self.options.tmpdir, self.mode),
-                logger=create_cfg.logger,
-                cluster_name=create_cfg.cluster_name,
-                ip_addr=create_cfg.ip_addr,
-                seeds=create_cfg.seeds,
-                cmdline_options=cmdline_options,
-                config_options=config_options,
-                property_file=create_cfg.property_file,
-                append_env=self.base_env,
-                server_encryption=create_cfg.server_encryption)
-
-            return server
-
-        async def create_cluster(logger: Union[logging.Logger, logging.LoggerAdapter]) -> ScyllaCluster:
-            cluster = ScyllaCluster(logger, self.hosts, cluster_size, create_server)
-
-            async def stop() -> None:
-                await cluster.stop()
-
-            # Suite artifacts are removed when
-            # the entire suite ends successfully.
-            self.artifacts.add_suite_artifact(self, stop)
-            if not self.options.save_log_on_success:
-                # If a test fails, we might want to keep the data dirs.
-                async def uninstall() -> None:
-                    await cluster.uninstall()
-
-                self.artifacts.add_suite_artifact(self, uninstall)
-            self.artifacts.add_exit_artifact(self, stop)
-
-            await cluster.install_and_start()
-            return cluster
-
-        return create_cluster
-
-    @property
-    def pattern(self) -> str:
-        return ["*_test.py","test_*.py"]
-
-    async def add_test(self, shortname, casename) -> None:
-        test = PythonTest(self.next_id((shortname, self.suite_key)), shortname, casename, self)
-        self.tests.append(test)
-
-    async def run(self, test: 'Test', options: argparse.Namespace):
-        if not os.access(self.scylla_exe, os.F_OK):
-            raise FileNotFoundError(f"{self.scylla_exe} does not exist.")
-        if not os.access(self.scylla_exe, os.X_OK):
-            raise PermissionError(f"{self.scylla_exe} is not executable.")
-        return await super().run(test, options)
-
-
-class CQLApprovalTestSuite(PythonTestSuite):
-    """Run CQL commands against a single Scylla instance"""
-
-    def __init__(self, path, cfg, options: argparse.Namespace, mode) -> None:
-        super().__init__(path, cfg, options, mode)
-
-    async def add_test(self, shortname: str, casename: str) -> None:
-        test = CQLApprovalTest(self.next_id((shortname, self.suite_key)), shortname, self)
-        self.tests.append(test)
-
-    @property
-    def pattern(self) -> str:
-        return "*test.cql"
-
-
-class TopologyTestSuite(PythonTestSuite):
-    """A collection of Python pytests against Scylla instances dealing with topology changes.
-       Instead of using a single Scylla cluster directly, there is a cluster manager handling
-       the lifecycle of clusters and bringing up new ones as needed. The cluster health checks
-       are done per test case.
-    """
-
-    async def add_test(self, shortname: str, casename: str) -> None:
-        """Add test to suite"""
-        test = TopologyTest(self.next_id((shortname, 'topology', self.mode)), shortname, casename, self)
-        self.tests.append(test)
-
-    def junit_tests(self):
-        """Return an empty list, since topology tests are excluded from an aggregated Junit report to prevent double
-        count in the CI report"""
-        return []
-
-
-class RunTestSuite(TestSuite):
-    """TestSuite for test directory with a 'run' script """
-
-    def __init__(self, path: str, cfg, options: argparse.Namespace, mode: str) -> None:
-        super().__init__(path, cfg, options, mode)
-        self.scylla_exe = path_to(self.mode, "scylla")
-        self.scylla_env = dict(self.base_env)
-        if self.mode == "coverage":
-            self.scylla_env = coverage.env(self.scylla_exe, distinct_id=self.name)
-
-        self.scylla_env['SCYLLA'] = self.scylla_exe
-
-    async def add_test(self, shortname, casename) -> None:
-        test = RunTest(self.next_id((shortname, self.suite_key)), shortname, self)
-        self.tests.append(test)
-
-    @property
-    def pattern(self) -> str:
-        return "run"
-
-
-class ToolTestSuite(TestSuite):
-    """A collection of Python pytests that test tools
-
-    These tests do not need an cluster setup for them. They invoke scylla
-    manually, in tool mode.
-    """
-
-    def __init__(self, path, cfg: dict, options: argparse.Namespace, mode: str) -> None:
-        super().__init__(path, cfg, options, mode)
-
-    @property
-    def pattern(self) -> str:
-        return ["*_test.py", "test_*.py"]
-
-    async def add_test(self, shortname, casename) -> None:
-        test = ToolTest(self.next_id((shortname, self.suite_key)), shortname, self)
-        self.tests.append(test)
-
-
-class Test:
-    """Base class for CQL, Unit and Boost tests"""
-    def __init__(self, test_no: int, shortname: str, suite) -> None:
-        self.id = test_no
-        self.path = ""
-        self.args: List[str] = []
-        # Arguments which are required by a program regardless of additional test specific arguments
-        self.core_args : List[str] = []
-        self.valid_exit_codes = [0]
-        # Name with test suite name
-        self.name = os.path.join(suite.name, shortname.split('.')[0])
-        # Name within the suite
-        self.shortname = shortname
-        self.mode = suite.mode
-        self.suite = suite
-        self.allure_dir = pathlib.Path(suite.options.tmpdir) / self.mode / 'allure'
-        # Unique file name, which is also readable by human, as filename prefix
-        self.uname = "{}.{}.{}".format(self.suite.name, self.shortname, self.id)
-        self.log_filename = pathlib.Path(suite.options.tmpdir) / self.mode / (self.uname + ".log")
-        self.log_filename.parent.mkdir(parents=True, exist_ok=True)
-        self.is_flaky = self.shortname in suite.flaky_tests
-        # True if the test was retried after it failed
-        self.is_flaky_failure = False
-        # True if the test was cancelled by a ctrl-c or timeout, so
-        # shouldn't be retried, even if it is flaky
-        self.is_cancelled = False
-        self.env = dict(self.suite.base_env)
-        self.started = False
-        self.success = False
-        self.time_start: float = 0
-        self.time_end: float = 0
-
-    def reset(self) -> None:
-        """Reset the test before a retry, if it is retried as flaky"""
-        self.success = False
-        self.time_start = 0
-        self.time_end = 0
-
-    @property
-    def failed(self):
-        """Returns True, if this test Failed"""
-        return self.started and not self.success and not self.is_cancelled
-
-    @property
-    def did_not_run(self):
-        """Returns True, if this test did not run correctly, i.e. was canceled either during or before execution"""
-        return not self.started or self.is_cancelled
-
-    @abstractmethod
-    async def run(self, options: argparse.Namespace) -> 'Test':
-        pass
-
-    @abstractmethod
-    def print_summary(self) -> None:
-        pass
-
-    async def setup(self, port, options):
-        """
-        Performs any necessary setup steps before running a test.
-        Returns (fn, txt, test_env) where:
-        fn  - is a cleanup function to call unconditionally after the test stops running
-        txt - is failure-injection description.
-        test_env - is a dictionary containing environment variables map specific for the test
-        """
-        return (lambda: 0, None,{})
-
-    def check_log(self, trim: bool) -> None:
-        """Check and trim logs and xml output for tests which have it"""
-        if trim:
-            self.log_filename.unlink()
-        pass
-
-    def write_junit_failure_report(self, xml_res: ET.Element) -> None:
-        assert not self.success
-        xml_fail = ET.SubElement(xml_res, 'failure')
-        xml_fail.text = "Test {} {} failed, check the log at {}".format(
-            self.path,
-            " ".join(self.args),
-            self.log_filename)
-        if self.log_filename.exists():
-            system_out = ET.SubElement(xml_res, 'system-out')
-            system_out.text = read_log(self.log_filename)
-
-
-class UnitTest(Test):
-    standard_args = shlex.split("--overprovisioned --unsafe-bypass-fsync 1 "
-                                "--kernel-page-cache 1 "
-                                "--blocked-reactor-notify-ms 2000000 --collectd 0 "
-                                "--max-networking-io-control-blocks=100 ")
-
-    def __init__(self, test_no: int, shortname: str, suite, args: str) -> None:
-        super().__init__(test_no, shortname, suite)
-        self.path = path_to(self.mode, "test", suite.name, shortname.split('.')[0])
-        self.args = shlex.split(args) + UnitTest.standard_args
-        if self.mode == "coverage":
-            self.env.update(coverage.env(self.path))
-
-    def print_summary(self) -> None:
-        print("Output of {} {}:".format(self.path, " ".join(self.args)))
-        print(read_log(self.log_filename))
-
-    async def run(self, options) -> Test:
-        self.success = await run_test(self, options, env=self.env)
-        logging.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-
-TestPath = collections.namedtuple('TestPath', ['suite_name', 'test_name', 'case_name'])
-
-class BoostTest(Test):
-    """A unit test which can produce its own XML output"""
-
-    standard_args = shlex.split("--overprovisioned --unsafe-bypass-fsync 1 "
-                                "--kernel-page-cache 1 "
-                                "--blocked-reactor-notify-ms 2000000 --collectd 0 "
-                                "--max-networking-io-control-blocks=100 ")
-
-    def __init__(self, test_no: int, shortname: str, suite, args: str,
-                 casename: Optional[str], allows_compaction_groups : bool, execname: Optional[str]) -> None:
-        boost_args = []
-        combined_test = True if execname else False
-        _shortname = shortname
-        if casename:
-            shortname += '.' + casename
-            if combined_test:
-                boost_args += ['--run_test=' + _shortname + '/' + casename]
-            else:
-                boost_args += ['--run_test=' + casename]
-        else:
-            if combined_test:
-                boost_args += ['--run_test=' + _shortname]
-
-        super().__init__(test_no, shortname, suite)
-        if combined_test:
-            self.path = path_to(self.mode, "test", suite.name, execname)
-        else:
-            self.path = path_to(self.mode, "test", suite.name, shortname.split('.')[0])
-        self.args = shlex.split(args) + UnitTest.standard_args
-        if self.mode == "coverage":
-            self.env.update(coverage.env(self.path))
-
-        self.xmlout = os.path.join(suite.options.tmpdir, self.mode, "xml", self.uname + ".xunit.xml")
-        boost_args += ['--report_level=no',
-                       '--logger=HRF,test_suite:XML,test_suite,' + self.xmlout]
-        boost_args += ['--catch_system_errors=no']  # causes undebuggable cores
-        boost_args += ['--color_output=false']
-        boost_args += ['--']
-        self.args = boost_args + self.args
-        self.casename = casename
-        self.__test_case_elements: list[ET.Element] = []
-        self.allows_compaction_groups = allows_compaction_groups
-
-    def reset(self) -> None:
-        """Reset the test before a retry, if it is retried as flaky"""
-        super().reset()
-        self.__test_case_elements = []
-
-    def get_test_cases(self) -> list[ET.Element]:
-        if not self.__test_case_elements:
-            self.__parse_logger()
-        return self.__test_case_elements
-
-    @staticmethod
-    def test_path_of_element(test: ET.Element) -> TestPath:
-        path = test.attrib['path']
-        prefix, case_name = path.rsplit('::', 1)
-        suite_name, test_name = prefix.split('.', 1)
-        return TestPath(suite_name, test_name, case_name)
-
-    def __parse_logger(self) -> None:
-        def attach_path_and_mode(test):
-            # attach the "path" to the test so we can group the tests by this string
-            test_name = test.attrib['name']
-            prefix = self.name.replace(os.path.sep, '.')
-            test.attrib['path'] = f'{prefix}::{test_name}'
-            test.attrib['mode'] = self.mode
-            return test
-
-        try:
-            root = ET.parse(self.xmlout).getroot()
-            # only keep the tests which actually ran, the skipped ones do not have
-            # TestingTime tag in the corresponding TestCase tag.
-            self.__test_case_elements = map(attach_path_and_mode,
-                                            root.findall(".//TestCase[TestingTime]"))
-            os.unlink(self.xmlout)
-        except ET.ParseError as e:
-            message = palette.crit(f"failed to parse XML output '{self.xmlout}': {e}")
-            if e.msg.__contains__("no element found"):
-                message = palette.crit(f"Empty testcase XML output, possibly caused by a crash in the cql_test_env.cc, "
-                                       f"details: '{self.xmlout}': {e}")
-            print(f"error: {self.name}: {message}")
-
-    def check_log(self, trim: bool) -> None:
-        self.__parse_logger()
-        super().check_log(trim)
-
-    async def run(self, options):
-        if options.random_seed:
-            self.args += ['--random-seed', options.random_seed]
-        if self.allows_compaction_groups and options.x_log2_compaction_groups:
-            self.args += [ "--x-log2-compaction-groups", str(options.x_log2_compaction_groups) ]
-        self.success = await run_test(self, options, env=self.env)
-        logging.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-    def write_junit_failure_report(self, xml_res: ET.Element) -> None:
-        """Does not write junit report for Jenkins legacy reasons"""
-        assert False
-
-    def print_summary(self) -> None:
-        print("Output of {} {}:".format(self.path, " ".join(self.args)))
-        print(read_log(self.log_filename))
-
-
-class LdapTest(BoostTest):
-    """A unit test which can produce its own XML output, and needs an ldap server"""
-
-    def __init__(self, test_no, shortname, args, suite):
-        super().__init__(test_no, shortname, args, suite, None, False, None)
-
-    async def setup(self, port, options):
-        instances_root = pathlib.Path(options.tmpdir) / self.mode / 'ldap_instances'
-        byte_limit = options.byte_limit if options.byte_limit else randint(0, 2000)
-        project_root = pathlib.Path(os.path.dirname(__file__))
-        return setup(project_root=project_root, port=port, instance_root=instances_root, byte_limit=byte_limit)
-
-
-class CQLApprovalTest(Test):
-    """Run a sequence of CQL commands against a standalone Scylla"""
-
-    def __init__(self, test_no: int, shortname: str, suite) -> None:
-        super().__init__(test_no, shortname, suite)
-        # Path to cql_repl driver, in the given build mode
-        self.path = "pytest"
-        self.cql = suite.suite_path / (self.shortname + ".cql")
-        self.result = suite.suite_path / (self.shortname + ".result")
-        self.tmpfile = os.path.join(suite.options.tmpdir, self.mode, self.uname + ".reject")
-        self.reject = suite.suite_path / (self.shortname + ".reject")
-        self.server_log: Optional[str] = None
-        self.server_log_filename: Optional[pathlib.Path] = None
-        self.is_before_test_ok = False
-        self.is_executed_ok = False
-        self.is_new = False
-        self.is_after_test_ok = False
-        self.is_equal_result = False
-        self.summary = "not run"
-        self.unidiff: Optional[str] = None
-        self.server_log = None
-        self.server_log_filename = None
-        self.env: Dict[str, str] = dict()
-        self._prepare_args(suite.options)
-
-    def reset(self) -> None:
-        """Reset the test before a retry, if it is retried as flaky"""
-        super().reset()
-        self.is_before_test_ok = False
-        self.is_executed_ok = False
-        self.is_new = False
-        self.is_after_test_ok = False
-        self.is_equal_result = False
-        self.summary = "not run"
-        self.unidiff = None
-        self.server_log = None
-        self.server_log_filename = None
-        self.env = dict()
-        old_tmpfile = pathlib.Path(self.tmpfile)
-        if old_tmpfile.exists():
-            old_tmpfile.unlink()
-
-
-    def _prepare_args(self, options: argparse.Namespace):
-        self.args = [
-            "-s",  # don't capture print() inside pytest
-            "test/pylib/cql_repl/cql_repl.py",
-            "--input={}".format(self.cql),
-            "--output={}".format(self.tmpfile),
-            "--run_id={}".format(self.id),
-            "--mode={}".format(self.mode),
-        ]
-        self.args.append(f"--alluredir={self.allure_dir}")
-        if not options.save_log_on_success:
-            self.args.append("--allure-no-capture")
-
-    async def run(self, options: argparse.Namespace) -> Test:
-        self.success = False
-        self.summary = "failed"
-
-        loggerPrefix = self.mode + '/' + self.uname
-        logger = LogPrefixAdapter(logging.getLogger(loggerPrefix), {'prefix': loggerPrefix})
-        def set_summary(summary):
-            self.summary = summary
-            log_func = logger.info if self.success else logger.error
-            log_func("Test %s %s", self.uname, summary)
-            if self.server_log is not None:
-                logger.info("Server log:\n%s", self.server_log)
-
-        # TODO: consider dirty_on_exception=True
-        async with (cm := self.suite.clusters.instance(False, logger)) as cluster:
-            try:
-                cluster.before_test(self.uname)
-                logger.info("Leasing Scylla cluster %s for test %s", cluster, self.uname)
-                self.args.insert(1, "--host={}".format(cluster.endpoint()))
-                # If pre-check fails, e.g. because Scylla failed to start
-                # or crashed between two tests, fail entire test.py
-                self.is_before_test_ok = True
-                cluster.take_log_savepoint()
-                self.is_executed_ok = await run_test(self, options, env=self.env)
-                cluster.after_test(self.uname, self.is_executed_ok)
-                cm.dirty = cluster.is_dirty
-                self.is_after_test_ok = True
-
-                if not self.is_executed_ok:
-                    set_summary("""returned non-zero return status.\n
-Check test log at {}.""".format(self.log_filename))
-                elif not os.path.isfile(self.tmpfile):
-                    set_summary("failed: no output file")
-                elif not os.path.isfile(self.result):
-                    set_summary("failed: no result file")
-                    self.is_new = True
-                else:
-                    self.is_equal_result = filecmp.cmp(self.result, self.tmpfile)
-                    if not self.is_equal_result:
-                        self.unidiff = format_unidiff(str(self.result), self.tmpfile)
-                        set_summary("failed: test output does not match expected result")
-                        assert self.unidiff is not None
-                        logger.info("\n{}".format(palette.nocolor(self.unidiff)))
-                    else:
-                        self.success = True
-                        set_summary("succeeded")
-            except Exception as e:
-                # Server log bloats the output if we produce it in all
-                # cases. So only grab it when it's relevant:
-                # 1) failed pre-check, e.g. start failure
-                # 2) failed test execution.
-                if not self.is_executed_ok:
-                    self.server_log = cluster.read_server_log()
-                    self.server_log_filename = cluster.server_log_filename()
-                    if not self.is_before_test_ok:
-                        set_summary("pre-check failed: {}".format(e))
-                        print("Test {} {}".format(self.name, self.summary))
-                        print("Server log  of the first server:\n{}".format(self.server_log))
-                        # Don't try to continue if the cluster is broken
-                        raise
-                set_summary("failed: {}".format(e))
-            finally:
-                if os.path.exists(self.tmpfile):
-                    if self.is_executed_ok and (self.is_new or not self.is_equal_result):
-                        # Move the .reject file close to the .result file
-                        # so that it's easy to analyze the diff or overwrite .result
-                        # with .reject.
-                        shutil.move(self.tmpfile, self.reject)
-                    else:
-                        pathlib.Path(self.tmpfile).unlink()
-
-        return self
-
-    def print_summary(self) -> None:
-        print("Test {} ({}) {}".format(palette.path(self.name), self.mode,
-                                       self.summary))
-        if not self.is_executed_ok:
-            print(read_log(self.log_filename))
-            if self.server_log is not None:
-                print("Server log of the first server:")
-                print(self.server_log)
-        elif not self.is_equal_result and self.unidiff:
-            print(self.unidiff)
-
-    def write_junit_failure_report(self, xml_res: ET.Element) -> None:
-        assert not self.success
-        xml_fail = ET.SubElement(xml_res, 'failure')
-        xml_fail.text = self.summary
-        if not self.is_executed_ok:
-            if self.log_filename.exists():
-                system_out = ET.SubElement(xml_res, 'system-out')
-                system_out.text = read_log(self.log_filename)
-            if self.server_log_filename:
-                system_err = ET.SubElement(xml_res, 'system-err')
-                system_err.text = read_log(self.server_log_filename)
-        elif self.unidiff:
-            system_out = ET.SubElement(xml_res, 'system-out')
-            system_out.text = palette.nocolor(self.unidiff)
-
-
-class RunTest(Test):
-    """Run tests in a directory started by a run script"""
-
-    def __init__(self, test_no: int, shortname: str, suite) -> None:
-        super().__init__(test_no, shortname, suite)
-        self.path = suite.suite_path / shortname
-        self.xmlout = os.path.join(suite.options.tmpdir, self.mode, "xml", self.uname + ".xunit.xml")
-        self.args = [
-            "--junit-xml={}".format(self.xmlout),
-            "-o",
-            "junit_suite_name={}".format(self.suite.name)
-        ]
-        self.args.append(f"--alluredir={self.allure_dir}")
-        if not suite.options.save_log_on_success:
-            self.args.append("--allure-no-capture")
-
-    def print_summary(self) -> None:
-        print("Output of {} {}:".format(self.path, " ".join(self.args)))
-        print(read_log(self.log_filename))
-
-    async def run(self, options: argparse.Namespace) -> Test:
-        # This test can and should be killed gently, with SIGTERM, not with SIGKILL
-        self.success = await run_test(self, options, gentle_kill=True, env=self.suite.scylla_env)
-        logging.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-
-class PythonTest(Test):
-    """Run a pytest collection of cases against a standalone Scylla"""
-
-    def __init__(self, test_no: int, shortname: str, casename: str, suite) -> None:
-        super().__init__(test_no, shortname, suite)
-        self.path = "python"
-        self.core_args = ["-m", "pytest"]
-        self.casename = casename
-        self.xmlout = os.path.join(self.suite.options.tmpdir, self.mode, "xml", self.uname + ".xunit.xml")
-        self.server_log: Optional[str] = None
-        self.server_log_filename: Optional[pathlib.Path] = None
-        self.is_before_test_ok = False
-        self.is_after_test_ok = False
-
-    def _prepare_pytest_params(self, options: argparse.Namespace):
-        self.args = [
-            "-s",  # don't capture print() output inside pytest
-            "--log-level=DEBUG",   # Capture logs
-            "-o",
-            "junit_family=xunit2",
-            "-o",
-            "junit_suite_name={}".format(self.suite.name),
-            "--junit-xml={}".format(self.xmlout),
-            "-rs",
-            "--run_id={}".format(self.id),
-            "--mode={}".format(self.mode)
-        ]
-        self.args.append(f"--alluredir={self.allure_dir}")
-        if not options.save_log_on_success:
-            self.args.append("--allure-no-capture")
-        if options.markers:
-            self.args.append(f"-m={options.markers}")
-
-            # https://docs.pytest.org/en/7.1.x/reference/exit-codes.html
-            no_tests_selected_exit_code = 5
-            self.valid_exit_codes = [0, no_tests_selected_exit_code]
-
-        arg = str(self.suite.suite_path / (self.shortname + ".py"))
-        if self.casename is not None:
-            arg += '::' + self.casename
-        self.args.append(arg)
-
-    def reset(self) -> None:
-        """Reset the test before a retry, if it is retried as flaky"""
-        super().reset()
-        self.server_log = None
-        self.server_log_filename = None
-        self.is_before_test_ok = False
-        self.is_after_test_ok = False
-
-    def print_summary(self) -> None:
-        print("Output of {} {}:".format(self.path, " ".join(self.args)))
-        print(read_log(self.log_filename))
-        if self.server_log is not None:
-            print("Server log of the first server:")
-            print(self.server_log)
-
-    async def run(self, options: argparse.Namespace) -> Test:
-
-        self._prepare_pytest_params(options)
-
-        loggerPrefix = self.mode + '/' + self.uname
-        logger = LogPrefixAdapter(logging.getLogger(loggerPrefix), {'prefix': loggerPrefix})
-        cluster = await self.suite.clusters.get(logger)
-        try:
-            cluster.before_test(self.uname)
-            prepare_cql = self.suite.cfg.get("prepare_cql", None)
-            if prepare_cql and not hasattr(cluster, 'prepare_cql_executed'):
-                cc = next(iter(cluster.running.values())).control_connection
-                if not isinstance(prepare_cql, collections.abc.Iterable):
-                    prepare_cql = [prepare_cql]
-                for stmt in prepare_cql:
-                    cc.execute(stmt)
-                cluster.prepare_cql_executed = True
-            logger.info("Leasing Scylla cluster %s for test %s", cluster, self.uname)
-            self.args.insert(0, "--host={}".format(cluster.endpoint()))
-            self.is_before_test_ok = True
-            cluster.take_log_savepoint()
-            status = await run_test(self, options, env=self.suite.scylla_env)
-            # if self.shortname in self.suite.dirties_cluster:
-            cluster.is_dirty = True
-            cluster.after_test(self.uname, status)
-            self.is_after_test_ok = True
-            self.success = status
-        except Exception as e:
-            self.server_log = cluster.read_server_log()
-            self.server_log_filename = cluster.server_log_filename()
-            if not self.is_before_test_ok:
-                print("Test {} pre-check failed: {}".format(self.name, str(e)))
-                print("Server log of the first server:\n{}".format(self.server_log))
-                logger.info(f"Discarding cluster after failed start for test %s...", self.name)
-            elif not self.is_after_test_ok:
-                print("Test {} post-check failed: {}".format(self.name, str(e)))
-                print("Server log of the first server:\n{}".format(self.server_log))
-                logger.info(f"Discarding cluster after failed test %s...", self.name)
-        await self.suite.clusters.put(cluster, is_dirty=cluster.is_dirty)
-        logger.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-    def write_junit_failure_report(self, xml_res: ET.Element) -> None:
-        super().write_junit_failure_report(xml_res)
-        if self.server_log_filename is not None:
-            system_err = ET.SubElement(xml_res, 'system-err')
-            system_err.text = read_log(self.server_log_filename)
-
-
-class TopologyTest(PythonTest):
-    """Run a pytest collection of cases against Scylla clusters handling topology changes"""
-    status: bool
-
-    def __init__(self, test_no: int, shortname: str, casename: str, suite) -> None:
-        super().__init__(test_no, shortname, casename, suite)
-
-    async def run(self, options: argparse.Namespace) -> Test:
-
-        self._prepare_pytest_params(options)
-
-        test_path = os.path.join(self.suite.options.tmpdir, self.mode)
-        async with get_cluster_manager(self.uname, self.suite.clusters, test_path) as manager:
-            self.args.insert(0, "--tmpdir={}".format(options.tmpdir))
-            self.args.insert(0, "--manager-api={}".format(manager.sock_path))
-            if options.artifacts_dir_url:
-                self.args.insert(0, "--artifacts_dir_url={}".format(options.artifacts_dir_url))
-
-            try:
-                # Note: start manager here so cluster (and its logs) is available in case of failure
-                await manager.start()
-                self.success = await run_test(self, options)
-            except Exception as e:
-                self.server_log = manager.cluster.read_server_log()
-                self.server_log_filename = manager.cluster.server_log_filename()
-                if not manager.is_before_test_ok:
-                    print("Test {} pre-check failed: {}".format(self.name, str(e)))
-                    print("Server log of the first server:\n{}".format(self.server_log))
-                    # Don't try to continue if the cluster is broken
-                    raise
-            manager.logger.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-
-class ToolTest(Test):
-    """Run a collection of pytest test cases
-
-    That do not need a scylla cluster set-up for them."""
-
-    def __init__(self, test_no: int, shortname: str, suite) -> None:
-        super().__init__(test_no, shortname, suite)
-        launcher = self.suite.cfg.get("launcher", "pytest")
-        self.path = launcher.split(maxsplit=1)[0]
-        self.xmlout = os.path.join(self.suite.options.tmpdir, self.mode, "xml", self.uname + ".xunit.xml")
-
-    def _prepare_pytest_params(self, options: argparse.Namespace):
-        launcher = self.suite.cfg.get("launcher", "pytest")
-        self.args = launcher.split()[1:]
-        self.args += [
-            "-s",  # don't capture print() output inside pytest
-            "--log-level=DEBUG",   # Capture logs
-            "-o",
-            "junit_family=xunit2",
-            "--junit-xml={}".format(self.xmlout),
-            "--mode={}".format(self.mode),
-            "--run_id={}".format(self.id)
-        ]
-        self.args.append(f"--alluredir={self.allure_dir}")
-        if not options.save_log_on_success:
-            self.args.append("--allure-no-capture")
-        if options.markers:
-            self.args.append(f"-m={options.markers}")
-
-            # https://docs.pytest.org/en/7.1.x/reference/exit-codes.html
-            no_tests_selected_exit_code = 5
-            self.valid_exit_codes = [0, no_tests_selected_exit_code]
-        self.args.append(str(self.suite.suite_path / (self.shortname + ".py")))
-
-    def print_summary(self) -> None:
-        print("Output of {} {}:".format(self.path, " ".join(self.args)))
-
-    async def run(self, options: argparse.Namespace) -> Test:
-        self._prepare_pytest_params(options)
-
-        loggerPrefix = self.mode + '/' + self.uname
-        logger = LogPrefixAdapter(logging.getLogger(loggerPrefix), {'prefix': loggerPrefix})
-        self.success = await run_test(self, options)
-        logger.info("Test %s %s", self.uname, "succeeded" if self.success else "failed ")
-        return self
-
-    def write_junit_failure_report(self, xml_res: ET.Element) -> None:
-        super().write_junit_failure_report(xml_res)
 
 
 class TabularConsoleOutput:
@@ -1321,148 +99,6 @@ class TabularConsoleOutput:
         else:
             msg += " {:.2f}s".format(test.time_end - test.time_start)
             print(msg)
-
-toxiproxy_id_gen = 0
-
-async def run_test(test: Test, options: argparse.Namespace, gentle_kill=False, env=dict()) -> bool:
-    """Run test program, return True if success else False"""
-
-    with test.log_filename.open("wb") as log:
-        global toxiproxy_id_gen
-        toxiproxy_id = toxiproxy_id_gen
-        toxiproxy_id_gen += 1
-        ldap_port = 5000 + (toxiproxy_id * 3) % 55000
-        cleanup_fn = None
-        finject_desc = None
-        def report_error(error, failure_injection_desc = None):
-            msg = "=== TEST.PY SUMMARY START ===\n"
-            msg += "{}\n".format(error)
-            msg += "=== TEST.PY SUMMARY END ===\n"
-            if failure_injection_desc is not None:
-                msg += 'failure injection: {}'.format(failure_injection_desc)
-            log.write(msg.encode(encoding="UTF-8"))
-
-        try:
-            cleanup_fn, finject_desc, test_env = await test.setup(ldap_port, options)
-        except Exception as e:
-            report_error("Test setup failed ({})\n{}".format(str(e), traceback.format_exc()))
-            return False
-        process = None
-        stdout = None
-        logging.info("Starting test %s: %s %s", test.uname, test.path, " ".join(test.args))
-        UBSAN_OPTIONS = [
-            "halt_on_error=1",
-            "abort_on_error=1",
-            f"suppressions={os.getcwd()}/ubsan-suppressions.supp",
-            os.getenv("UBSAN_OPTIONS"),
-        ]
-        ASAN_OPTIONS = [
-            "disable_coredump=0",
-            "abort_on_error=1",
-            "detect_stack_use_after_return=1",
-            os.getenv("ASAN_OPTIONS"),
-        ]
-        ldap_instance_path = os.path.join(
-            os.path.abspath(os.path.join(options.tmpdir, test.mode, 'ldap_instances')),
-            str(ldap_port))
-        saslauthd_mux_path = os.path.join(ldap_instance_path, 'mux')
-        if options.manual_execution:
-            print('Please run the following shell command, then press <enter>:')
-            test_env_string = " ".join([f"{k}={v}" for k,v in test_env.items()])
-            print('{} {}'.format(
-                test_env_string, ' '.join([shlex.quote(e) for e in [test.path, *test.args]])))
-            input('-- press <enter> to continue --')
-            if cleanup_fn is not None:
-                cleanup_fn()
-            return True
-        try:
-            resource_gather = get_resource_gather(options.gather_metrics, test, options.tmpdir)
-            resource_gather.make_cgroup()
-            log.write("=== TEST.PY STARTING TEST {} ===\n".format(test.uname).encode(encoding="UTF-8"))
-            log.write("export UBSAN_OPTIONS='{}'\n".format(
-                ":".join(filter(None, UBSAN_OPTIONS))).encode(encoding="UTF-8"))
-            log.write("export ASAN_OPTIONS='{}'\n".format(
-                ":".join(filter(None, ASAN_OPTIONS))).encode(encoding="UTF-8"))
-            for k,v in test_env.items():
-                log.write(f"export {k}={v}\n".encode(encoding="UTF-8"))
-            log.write("{} {}\n".format(test.path, " ".join(test.args)).encode(encoding="UTF-8"))
-            log.write("=== TEST.PY TEST {} OUTPUT ===\n".format(test.uname).encode(encoding="UTF-8"))
-            log.flush()
-            test.time_start = time.time()
-            test.time_end = 0
-
-            path = test.path
-            args = test.core_args + test.args
-            if options.cpus:
-                path = 'taskset'
-                args = ['-c', options.cpus, test.path, *args]
-
-            test_running_event = asyncio.Event()
-            test_resource_watcher = resource_gather.cgroup_monitor(test_event=test_running_event)
-
-            test_env.update(
-                dict(os.environ,
-                         UBSAN_OPTIONS=":".join(filter(None, UBSAN_OPTIONS)),
-                         ASAN_OPTIONS=":".join(filter(None, ASAN_OPTIONS)),
-                         # TMPDIR env variable is used by any seastar/scylla
-                         # test for directory to store test temporary data.
-                         TMPDIR=os.path.join(options.tmpdir, test.mode),
-                         SCYLLA_TEST_ENV='yes',
-                         **env,
-                 )
-            )
-            process = await asyncio.create_subprocess_exec(
-                path, *args,
-                stderr=log,
-                stdout=log,
-                env=test_env,
-                preexec_fn=resource_gather.put_process_to_cgroup,
-            )
-            stdout, _ = await asyncio.wait_for(process.communicate(), options.timeout)
-            test_running_event.set()
-            test.time_end = time.time()
-
-            metrics = resource_gather.get_test_metrics()
-            try:
-                async with asyncio.timeout(2):
-                    await test_resource_watcher
-            except TimeoutError:
-                log.write(f'Metrics for {test.name} can be inaccurate, job reached timeout'.encode(encoding='UTF-8'))
-            finally:
-                resource_gather.remove_cgroup()
-
-            if process.returncode not in test.valid_exit_codes:
-                report_error('Test exited with code {code}\n'.format(code=process.returncode))
-                resource_gather.write_metrics_to_db(metrics)
-                return False
-            try:
-                test.check_log(not options.save_log_on_success)
-            except Exception as e:
-                print("")
-                print(test.name + ": " + palette.crit("failed to parse XML output: {}".format(e)))
-                resource_gather.write_metrics_to_db(metrics)
-                return False
-            resource_gather.write_metrics_to_db(metrics, True)
-            return True
-        except (asyncio.TimeoutError, asyncio.CancelledError) as e:
-            test.is_cancelled = True
-            if process is not None:
-                if gentle_kill:
-                    process.terminate()
-                else:
-                    process.kill()
-                stdout, _ = await process.communicate()
-            if isinstance(e, asyncio.TimeoutError):
-                report_error("Test timed out")
-            elif isinstance(e, asyncio.CancelledError):
-                print(test.shortname, end=" ")
-                report_error("Test was cancelled: the parent process is exiting")
-        except Exception as e:
-            report_error("Failed to run the test:\n{e}".format(e=e))
-        finally:
-            if cleanup_fn is not None:
-                cleanup_fn()
-    return False
 
 
 def setup_signal_handlers(loop, signaled) -> None:
@@ -1741,18 +377,6 @@ async def run_all_tests(signaled: asyncio.Event, options: argparse.Namespace) ->
     console.print_end_blurb()
 
 
-def read_log(log_filename: pathlib.Path) -> str:
-    """Intelligently read test log output"""
-    try:
-        with log_filename.open("r") as log:
-            msg = log.read()
-            return msg if len(msg) else "===Empty log output==="
-    except FileNotFoundError:
-        return "===Log {} not found===".format(log_filename)
-    except OSError as e:
-        return "===Error reading log {}===".format(e)
-
-
 def print_summary(failed_tests: List["Test"], cancelled_tests: int, options: argparse.Namespace) -> None:
     rusage = resource.getrusage(resource.RUSAGE_CHILDREN)
     cpu_used = rusage.ru_stime + rusage.ru_utime
@@ -1770,31 +394,6 @@ def print_summary(failed_tests: List["Test"], cancelled_tests: int, options: arg
             print(f"Summary: {len(failed_tests)} of the total {TestSuite.test_count()} tests failed, {cancelled_tests} cancelled")
         else:
             print(f"Summary: {len(failed_tests)} of the total {TestSuite.test_count()} tests failed")
-
-
-def format_unidiff(fromfile: str, tofile: str) -> str:
-    with open(fromfile, "r") as frm, open(tofile, "r") as to:
-        buf = StringIO()
-        diff = difflib.unified_diff(
-            frm.readlines(),
-            to.readlines(),
-            fromfile=fromfile,
-            tofile=tofile,
-            fromfiledate=time.ctime(os.stat(fromfile).st_mtime),
-            tofiledate=time.ctime(os.stat(tofile).st_mtime),
-            n=10)           # Number of context lines
-
-        for i, line in enumerate(diff):
-            if i > 60:
-                break
-            if line.startswith('+'):
-                line = palette.diff_in(line)
-            elif line.startswith('-'):
-                line = palette.diff_out(line)
-            elif line.startswith('@'):
-                line = palette.diff_mark(line)
-            buf.write(line)
-        return buf.getvalue()
 
 
 def write_junit_report(tmpdir: str, mode: str) -> None:
@@ -2065,6 +664,7 @@ async def main() -> int:
     #       to cooperate with git bisect's expectations
     return 0 if not failed_tests else 1
 
+
 async def process_coverage(options):
     total_processing_time = time.time()
     logger = LogPrefixAdapter(logging.getLogger("coverage"), {'prefix' : 'coverage'})
@@ -2307,9 +907,6 @@ async def process_coverage(options):
     logger.info(f"Done creating textual report. ({options.tmpdir}/test_coverage_report.txt)")
     total_processing_time = time.time() - total_processing_time
     ROOT_NODE.data.time = total_processing_time
-
-
-
 
     stats_str ="\n" + stats.show(stdout=False,
                                  data_property="asstring")
