+++ test/pylib/suite/__init__.py
@@ -1,0 +1,14 @@
+#
+# Copyright (C) 2025-present ScyllaDB
+#
+# SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.0
+#
+
+from test.pylib.suite.boost import BoostTestSuite
+from test.pylib.suite.cql_approval import CQLApprovalTestSuite
+from test.pylib.suite.ldap import LdapTestSuite
+from test.pylib.suite.python import PythonTestSuite
+from test.pylib.suite.run import RunTestSuite
+from test.pylib.suite.tool import ToolTestSuite
+from test.pylib.suite.topology import TopologyTestSuite
+from test.pylib.suite.unit import UnitTestSuite
